+++ src/Phan/Output/Printer/PylintPrinter.php
@@ -2,6 +2,7 @@
 
 namespace Phan\Output\Printer;
 
+use Phan\Config;
 use Phan\Issue;
 use Phan\IssueInstance;
 use Phan\Output\IssuePrinterInterface;
@@ -29,6 +30,10 @@ final class PylintPrinter implements IssuePrinterInterface
             self::getSeverityCode($instance),
             $message
         );
+        $column  = $instance->getColumn();
+        if ($column > 0 && !Config::getValue('hide_issue_column')) {
+            $line .= " (at column $column)";
+        }
         $suggestion = $instance->getSuggestionMessage();
         if ($suggestion) {
             $line .= " ($suggestion)";