+++ tests/plugin_test/src/081_use_return_value.php
@@ -2,7 +2,7 @@
 sprintf("%s is not used\n", 'return value');
 count([]);
 -count([]);  // should also warn
-try {
+try {  // should emit PhanPluginEmptyStatementTryBody
 } catch (Exception $e) {
     // Should warn - The message is usually used.
     // TODO: Make PhanPluginUseReturnValueInternalKnown also check for methods that override methods of Exception
@@ -15,4 +15,12 @@ try {
     var_export(-$e->getCode());
     // should warn
     -$e->getCode();
+} finally {  // should emit PhanPluginEmptyStatementTryFinally
+}
+try {
+    throw new RuntimeException();
+} catch (Exception $e) {
+    echo $e;
+} finally // should emit PhanPluginEmptyStatementTryFinally
+{
 }